json-parser: fix parsing non-string arrays
json-parser() works in a way that if it can parse the JSON data to a
compatible syslog-ng key-value type, it should do that, otherwise it
should save it as a json-literal.

A syslog-ng list is always a list of strings. An array of integers
cannot be stored as a syslog-ng list, so it should be stored as a json
literal, so format-json() can reproduce the original JSON. The current
implementation loses the type information of the non-string array's
elements.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>