logmsg: LM_VT_INTEGER should be reported as "int" type in user visible messages
There was a misalignment how we recognized and reported integer types,
we reported LM_VT_INTEGER as "integer" while we didn't accept this as
input.

This patch will add recognizing the "integer" value when converting from
user visible contexts. We will also recognize "integer" as a type hint
from now on.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>