correlation-context: add _clear() method
When an explicit trigger causes the correlation to be closed (e.g. the
trigger() criteria is met or a multi-line traceback sequence ends), it
makes sense to reuse the CorrelationContext instance, is it is quite
possible that we will need it again, right after the correlation finishes.

In order to make this possible, the CorrelationContext will only be destroyed
by the expire timer (which means that we were idle on that specific
context for timeout() seconds). In case this was an explicit end of the
context, we will just remove the data and keep the struct and its
key intact.

correlation_context_clear() facilitates clearing the context in place in such
a manner.

This should avoid having to allocate the context for a per-message basis.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>