contrib: Update vim syntax
The vim syntax was introduced 17 years ago for syslong-ng 1.6 but did
not get updated since then.  It lacks a lot of syntax highlighting.

Update the syntax to match more of the current syslog-ng language.  Move
the syntax in a vim directory that can be used with pathogen / plugged
and add an autoload file.

Signed-off-by: Romain Tartière <romain@blogreen.org>