syslog-format: fix null termination of SDATA param names
This fixes two instances of missed string null-termination in SDATA name
parsing.

The commit also adds error handling when reaching buffer boundaries.
Previously, either the \0 byte or part of the param name was omitted.

strncpy() may not null-terminate the destination when the source string
is longer, so those calls have been replaced with g_strlcpy().

Signed-off-by: László Várady <laszlo.varady@anno.io>