stats: add StatsByteCounter
StatsByteCounter allows representing "big" accumulating byte-based metrics
even on 32-bit architectures, where a simple stats-counter would overflow
after 4 GiB.

When 64-bit counters are available, they will be used.
If not, the counters lose precision but are able to represent larger
values.

The current implementation is not thread-safe, reimplement it using a CAS
loop when needed.

Signed-off-by: László Várady <laszlo.varady@anno.io>