logpipe: add internal flag
We can write drivers in SCLs. There are some cases, where it would
be nice for a LogPipe to know, that it is part of an SCL block,
so it could behave differently, like not creating stats counters,
because the SCL takes care of it in a different way.

This flag will store this information.

LogPipes can be created in the following ways:
  * initiated from the single definition of it in the cfg
  * cloned beacuse of a reference in the config
  * created by another pipe (like LogReader and LogWriter)

The following commits will implement these.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>