python: fix reference counting of generator function
We are parsing the function pointer a couple lines above with "O".

O (object) [PyObject *]
Store a Python object (without any conversion) in a C object pointer.
The C program thus receives the actual object that was passed.
The object’s reference count is not increased.
The pointer stored is not NULL.

Its reference count is not increased with the parsing, however we
decrease its reference in python_config_generator_free().
This causes a problem, when we are reloading and re-registering the
same function.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>