python: handle ImportErrors from the C code
This patch adds a helpful message in cases where syslog-ng is unable to
import a module to promote the use of virtualenvs.

In our own packages, virtualenvs are standard and are automatically
initialized during package installation.

In distro packages however, virtualenv is not enabled by default, as
the packager is relying on the distros own packages to fulfill the
requirements of syslog-ng's Python plugins.

The virtualenv could make sense even in these cases, however. If I was
writing a custom Python script as a part of my configuration, I may need
a PyPI package that is not available as a part of my distribution. In
such cases, initializing a virtualenv would make sense.

As a part of this change, syslog-ng will stop reporting an error
if the virtualenv activation fails. Those cases will just assume we are
using the distro case, reporting an error at a non-debug level is not the
right solution.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>