python-modules/hypr: allow syslog-ng to start even if the Python requests module is missing
Hypr registers a confgen plugin, implemented in Python, which depends
on the Python requests module. If that is not present (e.g. because we
are not using a virtualenv), syslog-ng startup would fail. This
patch will defer the error to the actual initialization, this way
syslog-ng would start in these cases and would only fail if Hypr is
actually used.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>