mainloop-call: fix hang caused by the previous fix for pending mainloop-calls
The previous change that made exiting worker thread wait for pending
mainloop-calls to complete unfortunately causes a hang because the
ivykis pool lock is held in the worker thread while it waits,
and if the mainloop thread tries to take that lock while a mainloop-call
from the exiting thread is still pending then these two threads deadlock.

Fix this by moving the mainloop-call TLS state from the worker's TLS
structure to a separate structure that the worker's TLS points to.
Then rather than waiting for a pending mainloop-call to complete when
a worker thread wants to exit, just hand off responsibility for freeing
the mainloop-call structure when the mainloop-call is complete from the
worker thread to the mainloop thread itself.

Signed-off-by: Chuck Silvers <chuq@chuq.com>