scl: move scl.conf to ${config_includedir}
It helps to run syslog-ng in containers, with minimizing the content
of the etc/ dir, which is mounted usually.

Because of our file include logic if etc/scl.conf exists,
it is loaded and share/syslog-ng/include/scl.conf is ignored.
This is suitable for an upgrade path, as with an upgrade the old one
will be used, but with new installments the new one will be used.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>