stats-aggregator: move timer handling from registry to StatsAggregator
We only have a single aggregator using the timer functionality (e.g.
maximum and average are not time based at the moment) and the timer callback
in the registry would iterate over the entire hash of stats aggregators
at every invocation (e.g. O(N) algorithm).

If we are adding independent timers for stats aggregators we will use
an optimized AVL tree and that also allows us to use timers for
different timespans.

I also think that the time based orphaning of aggregators is easier to
understand if the timers are in StatsAggregators.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>