mainloop: don't reload plugin candidates at reload
There was an effort made in plugin.c to avoid rescanning plugins multiple
times (for instance at SIGHUP), but this was broken sometime along the
way.

This patch restores that behaviour so that even if plugins are
removed from the disk we would happily continue at SIGHUP. This also makes
reloads cheaper (as we won't have to load potentially heavyweight
shared objects from disk).

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>