logqueue: make the backlog mandatory
Previously the LogQueue interface allowed an optional backlog, that is
you could disable the existance of the sent-but-not-acknowledged backlog,
in fact this was the default behaviour.

However all production code _always_ enabled the backlog and only our
testcases were using the alternative path.

Get rid off log_queue_set_use_backlog() function and all related code.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>