logthrdestdrv: add worker-partition-key() option
This option can be used to direct similar messages to the same worker.
Similarity is determined based on hashing the value of the
worker-partition-key() template specified by the user.

It can be used to retain ordering for a subset of messages, or to
accumulate batches of certain types of messages.

Signed-off-by: László Várady <laszlo.varady@anno.io>