!WIP logthrdestdrv: add queue-partition-key() option
This option is for maintain a separate queue for similar messages.

Similarity is determined based on the value of the queue-partition-key()
template specified by the user.

It can be used to accumulate big/complete batch of messages without the
need to create too many workers. Workers currently pick up queues in a
round-robin manner.

Signed-off-by: László Várady <laszlo.varady@anno.io>