python-modules: do not package s3 on <= 3.6
Python 3.6 does not support fres boto3 and botocore versions, also
compiling type hinted code (which is done for some reason during
setup.py) does not seem to work.

We tackle these 2 in the following ways:

In requirements.txt we allow installing older boto3 and botocore
versions. We should not install them at all, but filtering based
on the python version would be complex.

We skip adding the s3 module to the packages in setup.py for <= 3.6.

The only platform we support and has 3.6 is centos-7. We can remove
these workarounds when the centos-7 support is dropped (June 2024).

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>