tls-context: add ignore-validity-period option for TLS
In some cases you want to verify the ssl certificate authenticity
but don't care about the validity period. For example when working
in a system without an NTP server.
Currently there is no way to do this except for simply not verifying
the certificate at all.

This option disables the ssl validity period check,
but still does all other verification steps.

Signed-off-by: Cedric Arickx <cedric.arickx@gmail.com>