syslog-ng/main: add a --check-startup option
This new option will perform a complete initialization of syslog-ng and
then exit, indicating the success of the initialization. This can be used
when you need to check if opening network ports and the entire
initialization would work or not.

This does not work if you have a syslog-ng running in the background, as
in those cases that would keep the ports open, causing the new one
to unable to open them again.

The success is indicated by an exit code:

When initialization is successful:
$ syslog-ng --check-startup
$ echo $?
0

When it is not:
$ syslog-ng --check-startup
$ echo $?
2

The exit codes match syslog-ng exit codes, e.g. 1 is a syntax error, 2
is a higher level initialization error.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>