cfg-grammar.y: fix location information for blocks
syslog-ng was giving me invalid error traces whenever I was using blocks,
at least if the starting brace of the block was not on the same line
as the "block" keyword. E.g:

# brace same line, works correctly
block source whatever() {
};

# brace on a separate line, did not work correctly
block source whatever(
  foo(bar)
  baz(bat)
)
{
};

The fix is to record the location of the block content instead of the block
statement as a whole. This fixes up @line references that are emitted
as blocks get expanded.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>