afsocket: don't activate the listener for port 0
When the port is set to 0, we would allocate a dynamic port for
our listeners. This can happen when we explicitly set port to 0 or when
we use an unknown service name (as syslog-ng will happily resolve the
service from /etc/services).

This behaviour does not add too much value, syslog-ng would open a port
randomly.

This patch changes afsocket to avoid opening the listener in these cases.

The intent for this change is to be able to list source drivers in a
complex driver such as default-network-drivers() and still be able to
disable the constituent drivers just by setting their port to 0.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>