Make building examples optional
Issue #4496 reports an issue where building the example module
random-choice-generator fails with clang++ due to the following error:

../lib/atomic-gssize.h:118:10: error: cannot initialize a parameter of type 'gssize *' (aka 'long *') with an rvalue of type 'void *'
  118 |   return g_atomic_pointer_compare_and_exchange(&a->value, oldval, newval);
/usr/include/glib-2.0/glib/gatomic.h:258:44: note: expanded from macro 'g_atomic_pointer_compare_and_exchange'
  258 |     __atomic_compare_exchange_n ((atomic), (void *) (&(gapcae_oldval)), (newval), FALSE, __ATOMIC_SEQ_CST, __ATOMIC_SEQ_CST) ? TRUE : FALSE; \

with passing '--disable-cpp' being the recommended work-around. This
commit proposes an alternate work-around by updating the system so
building example modules is optional.

The value of this work-around is it provides a means of reducing
compile-time and work for builders who do not need to build the
example modules.