cfg-grammar: eliminate last_template variable
In order to recurse into our grammar we can't use global variables to track
objects we are parsing into.

This solution uses the $0 value for rule actions to pass arguments to that
specific rule, this time the LogTemplate instance we need to parse into.

That way the LogTemplate instance would be separate even if we are
recursing into the same parts of the grammar.

A similar solution could potentially be established to all similar cases,
using global variables in a recursive grammar is not right.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>