template: add "escape" setting to LogTemplateOptions as well
With this patch it becomes possible to enable escaping from both the
LogTemplate side (e.g.  when template-escape(yes) is present in a template
statement) and the LogTemplateOptions side (e.g.  with the destination
driver specific template-escape()) option.

Previously only LogWriter supplied this setting and it only worked in case
of an inline template.  With this change, this works with all drivers that
support the `template_option` grammar rule.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>