logmsg: add "predefined" handles
Add the possibility of adding a set of LM_V_XXXX handles that are allocated
at startup but not "static" in the sense that they are stored in the
NVTable as normal dynamic name-value pairs.

This mechanism replaces LOG_MSG_GET_VALUE_HANDLE_STATIC() macro and
is easier to read and avoids races.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>