add-contextual-data: delay the freeing of the literal template string
We are freeing the original template string that we've read from the
database, as it can consume a lot of memory with large databases.

Originally, we did this right after instantiating the template.
This is earlier than the time we report that the database record
was added.

By delaying the call to log_template_forget_template_string() we get the
proper value of the template when we report it in our log message.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>