correlation: Capture the end of a message with NLSTRING
When parsing a message, NLSTRING stops when it finds a `\n` char.  If
the message does not contain any `\n` char, the matching fail and the
message portion is not matched.

However, the documentation says that:
> For single-line messages, NLSTRING is equivalent with ANYSTRING

In order to match what the documentation say, we need to not return an
error when `\n` is not found, and instead capture the end of the
message.

Signed-off-by: Romain Tartière <romain@blogreen.org>