light: fix readline() called while file is being written
readline() returns if it reaches "\n" or EOF.

If File::wait_for_lines::find_lines_in_file() is called when only a
part of a line was written to the file, readline() returns because of
EOF, and only the first part of the line is processed/compared, and
the rest of the line is processed separately. In this case we should
try to read multiple times until "\n", join the line parts, and process
the joined line as a whole.

This PR implements that.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>