afsocket: bump recvmsg() auxiliary data buffer to 256 bytes
This patch should fix #4597 where syslog-ng starts to use invalid PID values
on 32 bit platforms.

The reason was that the control buffer was truncated at a different
location when using 32 bit over using 64 bit architectures. On 64 bit, the
entire SCM_CREDENTIALS control message was missing, so we didn't extract
the ${.unix.pid}, ${.unix.uid} and ${.unix.gid} values.

On 32 bit platforms the header of these control messages was still present,
while the payload was not, causing the PID value to be extracted from
an invalid location.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>