dbld: use home dir of the container user
This commit fixes packaging on debian buster, which broke with the
removal of gosu.

While creating a python venv, some cache files are to be written to the
$HOME dir, which is /root, and the current user does not have a write
permission to it.

With the gosu method the $HOME dir was still /root, but python did not
try to use /root/.cache for some reason. I tried to pinpoint the cause,
but could not find it. I also do not know why it only affects buster,
but I wasted too much time to investigate.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>