logthrsourcedrv: multi-worker support
This commit:
  * Changes the API to work on a Worker instead of a Driver.
  * Follows up the API changes at the users of the API.
  * Adds a new "workers()" option, which is opt-in for threaded source
    based drivers.
  * Starts multiple workers if set by the "self->num_workers" field.

Where it was easily applicable, I have implemented the multi worker
support:
  * opentelemetry()
  * random-choice-generator()
  * example-random-generator()

Where it would have needed more work, I have added an assert and a
comment to point out that although those drivers are using the
multi-worker API, they are not prepared to work with multiple workers:
  * python-source()
  * LogThreadedFetcher

This is a breaking API change, but if I remember correctly, we do not
have any threaded source driver based drivers in PE, only fetcher based.
The fetcher based drivers might need a small change because of the new
"workers" field that replaces "worker", but it is a trivial one.
(See the python fetcher code change.)

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>