bigquery:  Fixed the check of the GRPC switch (using the same method like we have in other GRPC modules)
Yes, I know, i do not like it either, i would like to bring back in an upcoming round all the module switches for all the C++ modules (regardless if it is a GRPC-dependent or simple C++ one), but in a separate run.

Also, please NOTE that we do not have per-module switches now, as we removed them earlier based on an agreement.
Also, we still/again have none matching settings in the cmake and autotools config files, I'd liket to sync them again in the above mentioned upcoming C++/GRPC module switch re-introduction round.

Signed-off-by: Hofi <hofione@gmail.com>