grpc: add DestDriverMetrics class to handle common metrics
Iimplementing this with composition in each gRPC related dest
driver is suboptimal.

The clean thing would be to have a common gRPC related dest driver,
which already does handles metrics, and the otel, bigquery and loki
drivers should descend from it. Unfortunately doing that would be a
huge task right now.

IMO this should suffice for now, but we should really clean up
the class hierarchy sometime later.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>