otel: add deadline to shutdown
Shutdown() waits for all the working threads to finish,
but they only finish, when _worker_wakeup() is called,
but that only gets called after Shutdown(), so we have
a deadlock.

Giving a deadline solves the issue:
> If all pending calls complete (and all their operations are
> retrieved by Next) before \a deadline expires, this finishes
> gracefully. Otherwise, forcefully cancel all pending calls associated
> with the server after \a deadline expires.

I have also removed the null-ing, as that frees the Server, which
we still want to use in the worker threads for submitting the
response.

Calling Shutdown() seems to be idempotent.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>