filterx: make storing virtual methods simpler
For some reason certain C++ compiler versions do not seem to be able to
find the function pointers in the struct.

E.g. on debian-buster and debian-bullseye:

In file included from ../../modules/grpc/otel/filterx/otel-resource.hpp:29,
                 from ../../modules/grpc/otel/filterx/otel-resource.cpp:23:
../../lib/filterx/filterx-object.h:67:5: error: too many initializers for ‘FilterXType’ {aka ‘_FilterXType’}
   67 |     }
      |     ^

I did not investigate why certain compilers can and certain cannot
handle this. We can do that, but I tried to save some time by just
making the struct simpler and it seems to work.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>