otel: support borrowing ArrayValue in Array
When we get() an ArrayValue of a Message, the respective created
FilterXObject should operate on the ArrayValue.

The current code creates a new ArrayValue every time, which is
incorrect.

Please note that this commit has an ugly workaround of dropping a const
qualifier in FilterXObjectGetter(). I believe that argument should not
be const. Please see the next commit, which changes that.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>