otel: pass mutable message in ProtobufField Get()
It seems logical, that when we want to get a certain field
of a Message, we do not want to modify the Message, at least
this is true for immutable values.

However, with the Array support there came a new use case:
When we get() an AnyValue field, that has an ArrayValue in it,
we want to return such a FilterXOtelArray, that later on is able
to modify the ArrayValue in the Message itself. To do this, it
needs to store the pointer for the Message's mutable ArrayValue
when we get() it, and for this we need to pass a mutable Message
for the get() function.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>