otel: support borrowing KeyValueList in KVList
For more details, see:
  otel: support borrowing ArrayValue in Array

This commit also changes the internal representation of the data from
KeyValueList to RepeatedPtrField<KeyValue>. This is done, because the
KVList class intends to handle both "repeated KeyValue attributes" and
"KeyValueList kvlist_value". Essentially, those are the same thing, and
the OTel proto definition uses both.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>