otel: create kvlist implicitly for AnyValues
This makes it easier to write filterx code, as the user does not have
to manually call otel_kvlist() whenever they need to start a new map.

The only side effect is that if a key cannot be found in a KVList,
instead of failing, now we create a KV for it, and set its value to
null, which can be transformed to KVList later.

Example code:
  $otel_log = otel_logrecord();

  $otel_log.body = otel_kvlist();
  $otel_log.body["key_0"] = "foo";
  $otel_log.body["key_1"] = otel_kvlist();
  $otel_log.body["key_1"]["key_0"] = "bar";
  # body == {"key_0": "foo", "key_1":{"key_0": "bar"}}

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>