filterx: implement dict and list literal generators
We don't have generic dicts or lists in filterx. We have dict and a list
interface, which can be implemented by anyone.

Literal dicts and lists do not make sense in this case, however with
a bit of trickery we can make them "generator"s, which use the dict
and list interfaces and operate on an actual implementation.

For this we introduce the merge "<<" operator, which on the left hand
side expects an object that is a dict or list implementation, and a
"generator" on the right hand side (in the form of literal dict and
list for now, but functions will be added soon). The merge operator
passes the "fillable" to the generator, which then fills it
appropriatery.

Please note that this is a shallow merge:
  * lists are always appended
  * dict merging works like python's dict update() function

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>