metrics-pipe: don't reset *matched to TRUE
The general model for the matched value passed along in LogPathOptions
is to initialize it to TRUE and set it to FALSE in case a path
becomes non-matching.

This should not be an actual bug, as we would not get the message in
queue() unless it is still TRUE, but no need to overwrite that value.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>