filterx: refactor literal generator
This is a pretty large refactor, with seemingly no point of having it.

We will have a number of complex generator rules in our grammar.
Currently the generator's fillable must be known before creating
the generator or before collecting the elements of the literal dicts
and lists. We will not have this luxury when we will introduce
the = operator for generators or when we will have to handle normal
and generator functions with the exact same tokens/rules, regarding the
grammar. I have a working branch where I prepare all the possible
generator related grammars, and I have had a lot of shift/reduce
conflicts when I used last_fillable.

This refactor delays the setting of fillable after the generator is
created. For this I had to create a generator class/interface which
gets implemented by all the generators.

Currently the only descendant is the literal generator (which will be
split to two in the next commit), but the generator function will also
be a descendant, and I think the regexp matcher will be the third one.

The side-effect of this commit is that we no longer use setattr() for
the literal generators, but set_subscript(), but I think we can live
with this.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>