filterx: add scope generations to avoid having to clone the scope
Instead of having to clone variables at every transition between filterx
scopes, add a "generation" counter which indicates if a previously
stored value is considered part of the scope or not.

At every transition we bump the generation counter of the scope (and at clone
we set it to 0). If our lookup returns a variable from a previous
generation, it is considered unset, just as we would unset it if
we were to sync() or clone() the scope.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>