module: mqtt
The current MQTT destination implementation does not answer to MQTT ping
requests in time when only a low volume of messages are send to the MQTT
broker.

MQTT broker might close the connection because of the missing responses,
causing error messages in syslog-ng and even lost messages. One example is
the broker operated by AWS as part of their "IoT Core Broker" service.

This patch adds regular calls to MQTTClient_yield every 0.5 seconds. These
calls allow the MQTT client library to answer the ping requests.

The requirement for these calls is documented in the Paho MQTT C Client
Library documentation in the comparision between synchronous and
asynchronous clients:

"Client applications must call either MQTTClient_receive() or
MQTTClient_yield() relatively frequently in order to allow processing of
acknowledgements and the MQTT "pings" that keep the network connection to the
server alive."

Signed-off-by: Wolfram Joost <wjoost@users.noreply.github.com>