s3: only use sts calls when a role is set
Calling sts.get_caller_identity() will fail with endpoints that are
advertised as S3 compatible but provide no other AWS services, for
example Minio:

syslog-ng[80]: Exception while calling a Python function; caller='minio#0', class='syslogng.modules.s3.S3Destination', function='open', exception='ClientError: An error occurred (InvalidClientTokenId) when calling the GetCallerIdentity operation: The security token included in the request is invalid.'

This patch moves the call that's not used for anything but for
displaying a log message about the assumed identity, inside the
condition checking for a `role` option.

However, we can safely call get_caller_identity if we use/have sts
anyway, because according to AWS Docs, you don't need any permissions
for that. I'm not sure about the effectiveness and necessity of this
though.

> No permissions are required to perform this operation. If an administrator attaches a policy to your identity that explicitly denies access to the sts:GetCallerIdentity action, you can still perform this operation.

Signed-off-by: Mate Ory <mate.ory@axoflow.com>