filterx: make it possible for set_sub and setattr to realloc new_value
The interface currently work in the following way:
When a new_value is passed to for example set_subscript(),
the set_subscript() implementation must modify new_value
in such a way that represents the stored value.

This is (will be) important for our parser functions, as
it is (will be) a common pattern to create a list, pass
it to set_subscript(), then add some values to it later.
The end user won't be able to do this, as a cloning will
happen for them, but our parser codes need this so we can
write short and high performance logic.

Note: this is further detailed in this commit:
> filterx: rework setattr() and set_subscript() virtual fns

We worked well so far, as the type of new_value never changed,
we just replaced the inner data structured with the stored one.
(see the last lines in OtelKVListField::FilterXObjectSetter())

However, I am planning to introduce JSON <-> OTel conversion
through the dict and list interfaces, where for example new_value
starts as an OTel KVList, but ends up as a JSON object after
the set_subscript() call. For this, we need to release the
original otel_kvlist new_value and create a completely new
json_object new_value. For this we need to pass a pointer to it.

Please note that this commit also introduces a new argument
to the map_to_json() function: `FilterXObject **assoc_object`,
and a missing object caching in json_array's append().

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>