filterx: process stmt rules in the right order
With the original grammar, the last stmt was processed first and the
first one last, reverse that order so filterx_expr_set_location() is
called closer to the parsing of the stmt in question.

This means that the top of the include stack still contains the block we
were included from, thereby allowing us to extract the block content
that already contains the substituted data.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>