filterx: let assign/setattr/set-subscript return the rhs
Instead of always returning TRUE to indicate success, return the rhs
and introduce an "ignore_falsy_return" property on the expression to
indicate that a falsy return is to be ignored.

NULL returns are still considered errors. This removes the need for custom
trace messages to find the results of assignments.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>