filterx: move weakrefs from scope to eval context
This patch moves the weakrefs array from FilterXScope to FilterXEvalContext.
Also, all filterx blocks along the pipeline will share the same weakrefs
array. They are going to be freed as the entire pipeline finishes off.

This allows the use of the weakrefs mechanism even if we are outside of the
evaluation loop, but still assumes that all weakrefs are basically cached
until we finish off processing the message.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>