filterx: only sync filterx state at filter/parser/rewrite elements and drivers
This patch renames the previous PIF_SYNC_SCOPE to PIF_SYNC_FILTERX as it
better conveys the intent, sync the filterx state (e.g. variables) to
the message _before_ the LogPipe is executed.

I also changed the setup logic so that LogPipe instances do not have the
flag by default, and I explicitly set them for filter/parser/rewrite
elements and drivers (both source and destination).

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>