metrics: further extract metrics cache
The metrics TLS cache binds the lifecycle of the
counters to the lifecycle of the thread, which is
the same as the process's.

There are cases where a cache like this is useful,
but we want to bind the lifecycle of the counters
to a different entity, like a HTTP worker.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>
Signed-off-by: Hofi <hofione@gmail.com>