file-writer: fix partial write acknowledgements
If a partial write occurs, the written messages must be acknowledged.

It is important to note that a message may be written partially.
In that case, the message will be acknowledged once the complete write
is successful. If write errors occur during this process, the partially
written message may be duplicated.

Signed-off-by: László Várady <laszlo.varady@anno.io>